% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-football-data.R
\name{get_football_data}
\alias{get_football_data}
\title{Get season data for one league}
\usage{
get_football_data(season = 2017, league = "E0")
}
\arguments{
\item{season}{The year (as a number) that the required season began.
For example for the 2018/2019 season, use \code{2018}}

\item{league}{The code (as a string) for the required league. At the moment only \code{"E0"}
for the English Premier League is supported}
}
\description{
Retrieve season level data from \code{football-data.co.uk} in a tidy format.
The data is downloaded on a competition-season level
}
\examples{
\dontrun{
get_football_data(season = 2017, league = "E0")
}
}
