% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_to_team_level.R
\name{transform_to_team_level}
\alias{transform_to_team_level}
\title{Transform clean data from match to team level data}
\usage{
transform_to_team_level(.data)
}
\arguments{
\item{.data}{A result from a \code{clean_football_data}}
}
\description{
Retrieve season level data from \url{http://www.football-data.co.uk} in a tidy format.
The data is downloaded on a competition-season level
}
\examples{
\dontrun{
transform_to_team_level()
}
}
