% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_football_data.R
\name{clean_football_data}
\alias{clean_football_data}
\title{Make football data ready for R coding}
\usage{
clean_football_data(.data)
}
\arguments{
\item{.data}{The data that needs to be cleaned. Can also be used in a pi}
}
\description{
Takes data downloaded from \url{http://www.football-data.co.uk} using
 \code{\link[footballdatar]{get_football_data}} and cleans it
by turning all column names into snake case and formatting the date columns.
Pipe ready
}
\examples{
\dontrun{
clean_football_data(.data = football_data)
}
}
